[DCS][TOF] Make the data point random generator more reusable.
