+++ Detectors/DCS/testWorkflow/src/dcs-random-data-workflow.cxx
@@ -8,28 +8,28 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "DetectorsDCS/DataPointIdentifier.h"
-#include "DetectorsDCS/DataPointValue.h"
-#include "Framework/TypeTraits.h"
-#include <unordered_map>
-namespace o2::framework
-{
-template <>
-struct has_root_dictionary<std::unordered_map<o2::dcs::DataPointIdentifier, o2::dcs::DataPointValue>, void> : std::true_type {
-};
-} // namespace o2::framework
+// #include "DetectorsDCS/DataPointIdentifier.h"
+// #include "DetectorsDCS/DataPointValue.h"
+// #include "Framework/TypeTraits.h"
+// #include <unordered_map>
+// namespace o2::framework
+// {
+// template <>
+// struct has_root_dictionary<std::unordered_map<o2::dcs::DataPointIdentifier, o2::dcs::DataPointValue>, void> : std::true_type {
+// };
+// } // namespace o2::framework
 #include "Framework/DataProcessorSpec.h"
-#include "DCSDataGeneratorSpec.h"
+#include "DCStestWorkflow/DCSRandomDataGeneratorSpec.h"
 #include "DCSDataProcessorSpec.h"
 
 using namespace o2::framework;
 
-// we need to add workflow options before including Framework/runDataProcessing
-void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
-{
-  // option allowing to set parameters
-}
-
+// // we need to add workflow options before including Framework/runDataProcessing
+// void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
+// {
+//   // option allowing to set parameters
+// }
+//
 // ------------------------------------------------------------------
 
 #include "Framework/runDataProcessing.h"
@@ -37,7 +37,7 @@ void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
 WorkflowSpec defineDataProcessing(ConfigContext const& configcontext)
 {
   WorkflowSpec specs;
-  specs.emplace_back(getDCSDataGeneratorSpec());
+  specs.emplace_back(o2::dcs::test::getDCSRandomDataGeneratorSpec());
   specs.emplace_back(getDCSDataProcessorSpec());
   return specs;
 }
